chore: rename lib/deploy/ modules
