+++ packages/cli/src/lib/deploy/DeployJob.ts
@@ -1,4 +1,4 @@
-import { DeployStep, StepName, StepResult, UriOrPrevStepResult } from "./step";
+import { DeployStep, StepName, StepResult, UriOrPrevStepResult } from "./DeployStep";
 import { Logger } from "../logging";
 
 import { Uri } from "@polywrap/core-js";
