+++ packages/cli/src/lib/deploy/DeployModule.ts
@@ -1,0 +1,5 @@
+import { Uri } from "@polywrap/core-js";
+
+export interface DeployModule {
+  execute(uri: Uri, config?: unknown): Promise<Uri>;
+}
