+++ packages/cli/src/lib/deploy/DeployPackage.ts
@@ -1,0 +1,8 @@
+import { DeployModule } from "./DeployModule";
+
+import { Schema as JsonSchema } from "jsonschema";
+
+export interface DeployPackage {
+  deployModule: DeployModule;
+  manifestExt: JsonSchema | undefined;
+}
