+++ packages/cli/src/lib/deploy/index.ts
@@ -1,12 +1,5 @@
-import { DeployModule } from "./Deployer";
-
-import { Schema as JsonSchema } from "jsonschema";
-
 export * from "./Deployer";
-export * from "./step";
-export * from "./job";
-
-export interface DeployPackage {
-  deployModule: DeployModule;
-  manifestExt: JsonSchema | undefined;
-}
+export * from "./DeployModule";
+export * from "./DeployPackage";
+export * from "./DeployStep";
+export * from "./DeployJob";
